+++ src/Nova/Role.php
@@ -1,12 +1,14 @@
 <?php
 namespace Sereny\NovaPermissions\Nova;
 
+use Illuminate\Database\Query\Builder;
 use Laravel\Nova\Fields\ID;
 use Illuminate\Http\Request;
 use Laravel\Nova\Fields\Text;
 use Illuminate\Validation\Rule;
 use Laravel\Nova\Fields\Select;
 use Laravel\Nova\Fields\MorphToMany;
+use Laravel\Nova\Http\Requests\NovaRequest;
 use Sereny\NovaPermissions\Fields\Checkboxes;
 use Sereny\NovaPermissions\Models\Role as RoleModel;
 
@@ -50,6 +52,14 @@ class Role extends Resource
     public static $title = 'name';
 
     /**
+     * Indicates if the resource should be displayed in the sidebar.
+     *
+     */
+    public static $with = [
+        'permissions',
+    ];
+
+    /**
      * Get the fields displayed by the resource.
      *
      * @param  \Illuminate\Http\Request $request
@@ -88,11 +98,15 @@ class Role extends Resource
                         'label'  => __($permission->name),
                     ];
                 })
-                ->groupBy('group')
-                ->toArray()),
+                    ->groupBy('group')
+                    ->toArray()),
 
             Text::make(__('Users'), function () {
-                return $this->users()->count();
+                /**
+                 * We eager load count for the users relationship in the index query.
+                 * @see self::indexQuery()
+                 */
+                return isset($this->users_count) ? $this->users_count : $this->users()->count();
             })->exceptOnForms(),
 
             MorphToMany::make($userResource::label(), 'users', $userResource)
@@ -108,22 +122,36 @@ class Role extends Resource
         return __('Roles');
     }
 
-
     public static function singularLabel()
     {
         return __('Role');
     }
 
     /**
-     * Load all permissions
+     * Let's eager load the user count within the "index" query.
+     */
+    public static function indexQuery(NovaRequest $request, $query)
+    {
+        /** @var Builder|PermissionModel $query */
+        return parent::indexQuery($request, $query)->withCount('users');
+    }
+
+    /**
+     * Load all permissions and cache for 1 minute.
+     * Enough to avoid N+1 at the Role index page,
+     * and not long enough to have them stalled.
      *
      * @return \Illuminate\Database\Eloquent\Collection
      */
     protected function loadPermissions()
     {
-        /** @var class-string */
-        $permissionClass = config('permission.models.permission');
+        $expirationTime = config('permission.cache.nova_expiration_time', now()->addMinute());
 
-        return $permissionClass::all()->unique('name');
+        return cache()->remember('sereny-nova-permissions', $expirationTime, function () {
+            /** @var class-string */
+            $permissionClass = config('permission.models.permission');
+
+            return $permissionClass::all()->unique('name');
+        });
     }
 }